Add files via upload
This is my final project for the Google Python Crash course. I used two different ways to count the interesting words in a text file and create a Word Count image. 
The code shows my skills in using python modules, data structure, and visualization. The data structures include string, list, and dictionary.
It is a fun project to get started in learning python.